Add a HID device to g_multi
This adds a HID device to g_multi so that when the BBB boot sequence
modprobes g_multi, we get an additional keyboard on the host computer
that is connected to /dev/hidg0 (by default) on the BBB.  You can write
HID reports to that descriptor and get keypresses on the host.  The
original stuff like RNDIS network adapters still work too.

This is done mostly by combining the patterns in g_multi.c with what's
done in upstream's hid.c.

After this, on Windows device manager I see (when sorting by
connection):

    USB Root Hub (USB 3.0)
      USB Composite Device
        Gadget Serial (COM3)
        Linux USB Ethernet/RNDIS Gadget
        USB Input Device
          HID Keyboard Device
        USB Mass Storage Device
          Linux File-Stor Gadget USB Device
      USB Input Device
        HID-compliant vendor-defined device