Remove the kernel mods and fix README, LICENSE
Distributing just a kernel module isn't really possible because it
relies on functions from other modules that are buggy and would need to
be replaced.  I got it working by just modifying and building the entire
kernel and all modules - I'm going to post that to a separate repo.

Now that the kernel stuff is out, simplify README.md and all the code in
here is mine so I can license as MIT.